Structure mise a jour, la ligne est stockee
