+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/16 14:08:16 by cmariot           #+#    #+#             */
-/*   Updated: 2021/12/14 12:28:41 by cmariot          ###   ########.fr       */
+/*   Updated: 2021/12/14 14:18:12 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -80,16 +80,16 @@ typedef struct s_shell {
 
 // init_minishell.c
 void	init_minishell(t_shell *ministruct, char **env);
+void	free_minishell(t_shell *ministruct);
 
 // env.c
 t_env	*save_env(char **array);
-char	*get_env_value(char *name, t_env *env);
 
 // get_command.c
 void	get_command(t_shell *minishell, t_command_line *command_line);
 char	*get_prompt(t_shell *ministruct);
 
-// list_t_env.c
+// env_list_utils.c
 t_env	*ft_lstnew_env(void *name, char *value);
 void	ft_lstadd_back_env(t_env **alst, t_env *new);
 void	ft_lstdelone_env(t_env *env, void (*del)(void *));
@@ -97,9 +97,9 @@ void	ft_lstclear_env(t_env **env, void (*del)(void *));
 t_env	*ft_lstlast_env(t_env *lst);
 
 // env_builtin.c
-void	ft_env(t_env *env);
-void	ft_setenv(t_env *env, char *name, char *value);
-void	ft_unsetenv(t_env *env, char *name);
+void	env_builtin(t_env *env);
+void	setenv_builtin(t_env *env, char *name, char *value);
+void	unsetenv_builtin(t_env *env, char *name);
 
 // parse_line.c
 //int		parse_line(t_command_table *command, char *line);
