+++ srcs/env_list_utils.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   list_t_env.c                                       :+:      :+:    :+:   */
+/*   env_list_utils.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/13 08:32:45 by cmariot           #+#    #+#             */
-/*   Updated: 2021/12/13 08:56:09 by cmariot          ###   ########.fr       */
+/*   Updated: 2021/12/14 14:29:21 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
