+++ srcs/free_at_exit.c
@@ -1,34 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   setenv.c                                           :+:      :+:    :+:   */
+/*   free_at_exit.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/12/13 08:30:01 by cmariot           #+#    #+#             */
-/*   Updated: 2021/12/13 09:10:02 by cmariot          ###   ########.fr       */
+/*   Created: 2021/12/14 14:11:32 by cmariot           #+#    #+#             */
+/*   Updated: 2021/12/14 14:29:29 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-// If the element name is in the linked list env, change its value,
-// else add name and value
-void	ft_setenv(t_env *env, char *name, char *value)
+//// Free the structure elements before exit
+void	free_minishell(t_shell *minishell)
 {
-	t_env	*tmp;
-
-	tmp = env;
-	while (env)
-	{
-		if (ft_strcmp(env->name, name) == 0)
-		{
-			free(env->value);
-			env->value = ft_strdup(value);
-			return ;
-		}
-		env = env->next;
-	}
-	env = tmp;
-	ft_lstadd_back_env(&env, ft_lstnew_env(name, value));
+	ft_lstclear_env(&minishell->env, free);
+	free(minishell->prompt);
+	free(minishell->command_line.line);
 }
